Added changes w.r.t proposal and moved batching into seperate class
Fixed shouldRestartCombinedNodesInExpectedOrderAndBatchedWithUrp
Moved nodeIdToKafkaNode from Batching to Availibility class for just improvement

Added changes w.r.t to the secrets changes

Added javadocs and fixed checkstyle issues
