+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/resource/rolling/PlatformClientImpl.java
@@ -8,9 +8,9 @@ import io.fabric8.kubernetes.api.model.Pod;
 import io.strimzi.operator.cluster.model.NodeRef;
 import io.strimzi.operator.cluster.model.RestartReasons;
 import io.strimzi.operator.cluster.operator.resource.events.KubernetesRestartEventPublisher;
+import io.strimzi.operator.cluster.operator.resource.kubernetes.PodOperator;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.model.Labels;
-import io.strimzi.operator.common.operator.resource.PodOperator;
 
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
