+++ operator-common/src/main/java/io/strimzi/operator/common/AdminClientProvider.java
@@ -29,12 +29,11 @@ public interface AdminClientProvider {
      * Create a Kafka Admin interface instance for controllers
      *
      * @param controllerBootstrapHostnames Kafka controller hostname to connect to for administration operations
-     * @param clusterCaCertSecret Secret containing the cluster CA certificate for TLS encryption
-     * @param keyCertSecret Secret containing keystore for TLS client authentication
-     * @param keyCertName Key inside the keyCertSecret for getting the keystore and the corresponding password
+     * @param kafkaCaTrustSet Trust set for connecting to Kafka
+     * @param authIdentity Identity for TLS client authentication for connecting to Kafka
      * @return Instance of Kafka Admin interface
      */
-    Admin createControllerAdminClient(String controllerBootstrapHostnames, Secret clusterCaCertSecret, Secret keyCertSecret, String keyCertName);
+    Admin createControllerAdminClient(String controllerBootstrapHostnames, PemTrustSet kafkaCaTrustSet, PemAuthIdentity authIdentity);
 
     /**
      * Create a Kafka Admin interface instance
